Move peak, convex_hull and pnpoly to _shared
