+++ skimage/_shared/setup.py
@@ -17,7 +17,9 @@ def configuration(parent_package='', top_path=None):
             'interpolation.pyx',
             'fast_exp.pyx',
             'warp.pyx',
-            'label.pyx'], working_path=base_path)
+            'label.pyx',
+            'convex_hull.pyx',
+            'pnpoly.pyx'], working_path=base_path)
 
     config.add_extension('geometry', sources=['geometry.c'])
     config.add_extension('transform', sources=['transform.c'],
@@ -27,6 +29,10 @@ def configuration(parent_package='', top_path=None):
     config.add_extension('warp', sources=['warp.c'],
                          include_dirs=[get_numpy_include_dirs()])
     config.add_extension('label', sources=['label.c'],
+                         include_dirs=[get_numpy_include_dirs()])
+    config.add_extension('convex_hull', sources=['convex_hull.c'],
+                         include_dirs=[get_numpy_include_dirs()])
+    config.add_extension('pnpoly', sources=['pnpoly.c'],
                          include_dirs=[get_numpy_include_dirs()])
     return config
