+++ skimage/measure/_regionprops.py
@@ -12,7 +12,7 @@ from ._find_contours import find_contours
 from ._marching_cubes_lewiner import marching_cubes
 from ._regionprops_utils import euler_number, perimeter, perimeter_crofton
 
-from ..morphology.convex_hull import convex_hull_image
+from .._shared.convex_hull import convex_hull_image
 
 
 __all__ = ['regionprops', 'euler_number', 'perimeter', 'perimeter_crofton']
