+++ skimage/measure/setup.py
@@ -14,8 +14,7 @@ def configuration(parent_package='', top_path=None):
     cython(['_find_contours_cy.pyx',
             '_moments_cy.pyx',
             '_marching_cubes_classic_cy.pyx',
-            '_marching_cubes_lewiner_cy.pyx',
-            '_pnpoly.pyx'], working_path=base_path)
+            '_marching_cubes_lewiner_cy.pyx'], working_path=base_path)
 
     config.add_extension('_find_contours_cy', sources=['_find_contours_cy.c'],
                          include_dirs=[get_numpy_include_dirs()])
@@ -27,8 +26,6 @@ def configuration(parent_package='', top_path=None):
     config.add_extension('_marching_cubes_lewiner_cy',
                          sources=['_marching_cubes_lewiner_cy.c'],
                          include_dirs=[get_numpy_include_dirs()])
-    config.add_extension('_pnpoly', sources=['_pnpoly.c'],
-                         include_dirs=[get_numpy_include_dirs(), '../_shared'])
 
     return config
