+++ skimage/morphology/setup.py
@@ -12,7 +12,6 @@ def configuration(parent_package='', top_path=None):
     config = Configuration('morphology', parent_package, top_path)
 
     cython(['_skeletonize_cy.pyx',
-            '_convex_hull.pyx',
             '_grayreconstruct.pyx',
             '_extrema_cy.pyx'], working_path=base_path)
     # _skeletonize_3d uses c++, so it must be cythonized separately
@@ -22,8 +21,6 @@ def configuration(parent_package='', top_path=None):
     cython(['_max_tree.pyx'], working_path=base_path)
 
     config.add_extension('_skeletonize_cy', sources=['_skeletonize_cy.c'],
-                         include_dirs=[get_numpy_include_dirs()])
-    config.add_extension('_convex_hull', sources=['_convex_hull.c'],
                          include_dirs=[get_numpy_include_dirs()])
     config.add_extension('_grayreconstruct', sources=['_grayreconstruct.c'],
                          include_dirs=[get_numpy_include_dirs()])
