+++ skimage/morphology/tests/test_convex_hull.py
@@ -1,6 +1,6 @@
 import numpy as np
 from skimage.morphology import convex_hull_image, convex_hull_object
-from skimage.morphology._convex_hull import possible_hull
+from skimage._shared.convex_hull import possible_hull
 
 from skimage._shared import testing
 from skimage._shared.testing import assert_array_equal
